Remove dom related elements from bot
- Remove underscore

- Remove JS DOM
