+++ src/common/mochaHelper.js
@@ -1,3 +1,2 @@
 /* eslint-disable import/no-extraneous-dependencies */
-require('binary-common-utils/lib/mochaHelper')
 require('app-module-path/register')
